+++ Modules/Intent.Modules.AmazonS3.ObjectStorage/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.2
+
+- Improvement: Updated templates to use new NuGet package system.
+
 ### Version 1.0.1
 
 - Improvement: Updated NuGet packages to latest stables.