+++ Modules/Intent.Modules.Application.DependencyInjection.MediatR/release-notes.md
@@ -1,3 +1,7 @@
+### Version 3.5.14
+
+- Improvement: Updated templates to use new NuGet package system.
+
 ### Version 3.5.13
 
 - Improvement: Updated NuGet packages to latest stables.