+++ Modules/Intent.Modules.Application.DependencyInjection/release-notes.md
@@ -1,3 +1,7 @@
+### Version 4.1.2
+
+- Improvement: Updated templates to use new NuGet package system.
+
 ### Version 4.1.1
 
 - Improvement: Updated NuGet packages to latest stables.